refactor: move path-level metrics to kinematics/path.py and add compute_path_straightness
