+++ movement/kinematics/__init__.py
@@ -6,7 +6,6 @@ from movement.kinematics.kinematics import (
     compute_displacement,
     compute_forward_displacement,
     compute_backward_displacement,
-    compute_path_length,
     compute_speed,
     compute_time_derivative,
     compute_velocity,
@@ -18,8 +17,9 @@ from movement.kinematics.orientation import (
     compute_turning_angle,
 )
 from movement.kinematics.kinetic_energy import compute_kinetic_energy
-from movement.kinematics.trajectory_complexity import (
-    compute_straightness_index,
+from movement.kinematics.path import (
+    compute_path_length,
+    compute_path_straightness,
 )
