+++ tests/test_unit/test_path.py
@@ -1,10 +1,10 @@
-"""Tests for compute_straightness_index."""
+"""Tests for compute_path_straightness."""
 
 import numpy as np
 import pytest
 import xarray as xr
 
-from movement.kinematics import compute_straightness_index
+from movement.kinematics.path import compute_path_straightness
 
 # ─────────────────────────────────────────────
 # Test dataset factories
@@ -33,7 +33,7 @@ def _create_test_dataarray(positions, n_ind=1, n_kp=1):
 def make_straight_line(length=10, n_ind=1, n_kp=1):
     """Straight diagonal line — SI should be exactly 1.0."""
     t = np.arange(length, dtype=float)
-    positions = np.stack([t, t], axis=-1)  # diagonal
+    positions = np.stack([t, t], axis=-1)
     return _create_test_dataarray(positions, n_ind, n_kp)
 
 
@@ -45,7 +45,7 @@ def make_closed_loop(n_ind=1, n_kp=1):
             [1.0, 0.0],
             [1.0, 1.0],
             [0.0, 1.0],
-            [0.0, 0.0],  # back to start
+            [0.0, 0.0],
         ]
     )
     return _create_test_dataarray(positions, n_ind, n_kp)
@@ -85,38 +85,29 @@ def make_known_si(n_ind=1, n_kp=1):
 # ─────────────────────────────────────────────
 
 
-def test_straightness_straight_line_is_one():
-    """Perfectly straight line should have SI = 1.0."""
-    data = make_straight_line(length=10)
-    result = compute_straightness_index(data)
-    assert np.allclose(result.values, 1.0, atol=1e-10)
-
-
-def test_straightness_closed_loop_is_zero():
-    """Trajectory returning to start should have SI = 0.0."""
-    data = make_closed_loop()
-    result = compute_straightness_index(data)
-    assert np.allclose(result.values, 0.0, atol=1e-10)
-
-
-def test_straightness_known_value():
-    """L-shaped path should return SI = 5/7 ≈ 0.7143."""
-    data = make_known_si()
-    result = compute_straightness_index(data)
-    assert np.allclose(result.values, 5 / 7, atol=1e-6)
+@pytest.mark.parametrize(
+    "data, expected",
+    [
+        (make_straight_line(length=10), 1.0),
+        (make_closed_loop(), 0.0),
+        (make_known_si(), 5 / 7),
+    ],
+)
+def test_straightness_known_values(data, expected):
+    """SI returns correct value for known trajectories."""
+    result = compute_path_straightness(data)
+    assert np.allclose(result.values, expected, atol=1e-6)
 
 
 def test_straightness_stationary_is_nan():
     """Stationary animal has path length 0 → SI should be NaN."""
-    data = make_stationary()
-    result = compute_straightness_index(data)
+    result = compute_path_straightness(make_stationary())
     assert np.all(np.isnan(result.values))
 
 
 def test_straightness_between_zero_and_one():
     """SI must be in [0, 1] for any valid trajectory."""
-    data = make_known_si()
-    result = compute_straightness_index(data)
+    result = compute_path_straightness(make_known_si())
     assert np.all((result.values >= 0) & (result.values <= 1))
 
 
@@ -125,27 +116,19 @@ def test_straightness_between_zero_and_one():
 # ─────────────────────────────────────────────
 
 
-def test_straightness_nan_at_start():
-    """NaN at start position propagates to NaN SI."""
+@pytest.mark.parametrize(
+    "nan_slice",
+    [
+        (0, Ellipsis),  # NaN at start
+        (-1, Ellipsis),  # NaN at end
+        (slice(None), Ellipsis),  # all NaN
+    ],
+)
+def test_straightness_nan_propagates(nan_slice):
+    """NaN at start, end, or everywhere propagates to NaN SI."""
     data = make_straight_line(length=10)
-    data[0, ...] = np.nan
-    result = compute_straightness_index(data)
-    assert np.all(np.isnan(result.values))
-
-
-def test_straightness_nan_at_end():
-    """NaN at end position propagates to NaN SI."""
-    data = make_straight_line(length=10)
-    data[-1, ...] = np.nan
-    result = compute_straightness_index(data)
-    assert np.all(np.isnan(result.values))
-
-
-def test_straightness_all_nan():
-    """All-NaN input should return NaN without crashing."""
-    data = make_straight_line(length=10)
-    data[:, ...] = np.nan
-    result = compute_straightness_index(data)
+    data[nan_slice] = np.nan
+    result = compute_path_straightness(data)
     assert np.all(np.isnan(result.values))
 
 
@@ -154,75 +137,31 @@ def test_straightness_all_nan():
 # ─────────────────────────────────────────────
 
 
-def test_straightness_drops_space_and_time_dims():
-    """Global SI should drop both time and space dimensions."""
-    data = make_straight_line(length=10, n_ind=2, n_kp=3)
-    result = compute_straightness_index(data)
+@pytest.mark.parametrize(
+    "n_ind, n_kp",
+    [
+        (1, 1),
+        (2, 3),
+        (3, 2),
+    ],
+)
+def test_straightness_output_shape(n_ind, n_kp):
+    """SI drops time and space dims, preserves individuals and keypoints."""
+    data = make_straight_line(length=10, n_ind=n_ind, n_kp=n_kp)
+    result = compute_path_straightness(data)
     assert "time" not in result.dims
     assert "space" not in result.dims
     assert "individuals" in result.dims
     assert "keypoints" in result.dims
-    assert result.sizes["individuals"] == 2
-    assert result.sizes["keypoints"] == 3
+    assert result.sizes["individuals"] == n_ind
+    assert result.sizes["keypoints"] == n_kp
 
 
 def test_straightness_output_name_and_units():
-    """Output should have name and units attributes."""
-    data = make_straight_line()
-    result = compute_straightness_index(data)
+    """Output should have correct name and units attributes."""
+    result = compute_path_straightness(make_straight_line())
     assert result.name == "straightness_index"
     assert result.attrs.get("units") == "dimensionless"
-
-
-def test_straightness_multiple_individuals():
-    """SI computed correctly for multiple individuals simultaneously."""
-    data = make_straight_line(length=10, n_ind=3, n_kp=2)
-    result = compute_straightness_index(data)
-    assert result.sizes["individuals"] == 3
-    assert result.sizes["keypoints"] == 2
-    # All straight lines → SI = 1.0 for all
-    assert np.allclose(result.values, 1.0, atol=1e-10)
-
-
-# ─────────────────────────────────────────────
-# Rolling window tests
-# ─────────────────────────────────────────────
-
-
-def test_straightness_rolling_output_shape():
-    """Rolling SI should preserve time dimension."""
-    data = make_straight_line(length=20)
-    result = compute_straightness_index(data, window_size=5)
-    assert "time" in result.dims
-    assert result.sizes["time"] == data.sizes["time"]
-    assert "space" not in result.dims
-
-
-def test_straightness_rolling_first_values_nan():
-    """First window_size - 1 values should be NaN (incomplete window)."""
-    data = make_straight_line(length=20)
-    W = 5
-    result = compute_straightness_index(data, window_size=W)
-    # First W values should be NaN
-    assert np.all(np.isnan(result.isel(time=slice(0, W)).values))
-
-
-def test_straightness_rolling_straight_line_is_one():
-    """Rolling SI on straight line should be 1.0 for all complete windows."""
-    data = make_straight_line(length=20)
-    W = 5
-    result = compute_straightness_index(data, window_size=W)
-    valid = result.isel(time=slice(W, None))
-    assert np.allclose(valid.values, 1.0, atol=1e-6)
-
-
-def test_straightness_rolling_invalid_window_size():
-    """Negative or zero window_size should raise ValueError."""
-    data = make_straight_line()
-    with pytest.raises(ValueError):
-        compute_straightness_index(data, window_size=0)
-    with pytest.raises(ValueError):
-        compute_straightness_index(data, window_size=-5)
 
 
 # ─────────────────────────────────────────────
@@ -238,4 +177,11 @@ def test_straightness_rejects_missing_time_dim():
         coords={"space": ["x", "y"]},
     )
     with pytest.raises(ValueError):
-        compute_straightness_index(data)
+        compute_path_straightness(data)
+
+
+def test_straightness_rejects_too_short():
+    """Should raise ValueError when fewer than 2 time points."""
+    data = make_straight_line(length=1)
+    with pytest.raises(ValueError):
+        compute_path_straightness(data)
